requirements.txt
fastapi
uvicorn
openai>=1.0.0
googletrans-py
google-trans-new
pydantic
